requirements.txt
numpy
pandas
sqlalchemy
mysql.connector
streamlit
python-dotenv
google-generativeai
pymysql